requirements.txt
fastapi
uvicorn
pymongo
pydantic
pyjwt
python-decouple
cloudinary
python-multipart